CI: Fix GHA concurrency
The previous concurrency group made it such that a patch series pushed
at the same time would cancel all the runs on the earlier changes
thereby causing them to not get voted on.

Since each GERRIT_CHANGE_ID is unique and we should only be running a
single instance of a job at a time for that specific GERRIT_CHANGE_ID
this is a valid concurrency group identifier. Use the run_id as a
fallback.

Change-Id: I9b417c5fe288c4b70ad0a776c24e1a50e7526ef3
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>