CI: GHA linting and add actionlint binary
In preparation for adding in linting of GHA files themselves we need to
make sure that the actionlint binary is installed. A follow on change
will add the pre-commit hook, adding the hook before having the verify
job install the binary will cause the verify job itself to fail with no
way forward.

Change-Id: Ib4d0a168c9c0b34d8b19ca6ba2152d3deabc599b
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>