Refactor!: Only search workflow filenames for jobs
In preparation for creating a method for doing required workflows along
with the standard workflows the names of workflows need to be managed a
little tighter. Going forward only the workflow filenames will be
searched for the relevant keyword and the filename will also be required
to have 'gerrit' in the name. Workflow "names" will no longer be
searched for workflows to trigger.

Change-Id: I1dada55d382e4504794fdca5f90a51219b243d86
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>