Feat: Add get_magic_repo helper
GitHub has a concept of a "magic" repo (.github) where organization
level components are stored. GitLab, might have such a concept, but we
are far from implementing for GitLab so we'll just assume GitHub centric
design for now.

This magic repo will be used for storing required workflows for
triggering. In particular, this will allow us to have a single global
INFO.yaml validator that we can use.

Change-Id: I35b865a8e7cb91cd878e99506428f26713801450
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>