Fix: On error continue processing workflows
In the case where a workflow does not trigger for reason and throws an
exception, this stops processing of all other workflows that were
detected. This an cause problems as other workflows may pass. As there
is no easy way to bubble triggering issues to end users not having jobs
(that might pass) not trigger because of a failure of a different
workflow (that doesn't get reported) is a problem.

Issue: RELENG-5100
Change-Id: I0bd20efa6cc8077e7453d4523e40ee92052c825a
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>